cow and ponder working
