+++ cowponder_pip/src/cowponder/__init__.py
@@ -3,3 +3,5 @@ cowponder
 
 A simple terminal command that displays randomly selected philosophical thoughts from a cow
 """
+
+from cowponder.ponder import ponder, cowponder, add_thoughts, update_thoughtbook
